refactor: split transform manager types
